Moved the settings Implementation and Structure to settings.rs - JIRA-47
